requirements.txt
Flask[async]
python-telegram-bot
gunicorn
requests
uvicorn